+++ languagetool-language-modules/gl/src/main/java/org/languagetool/rules/gl/GalicianWordinessRule.java
@@ -27,8 +27,6 @@ import org.languagetool.rules.ITSIssueType;
 import java.io.IOException;
 import java.util.Locale;
 import java.util.ResourceBundle;
-import java.net.MalformedURLException;
-import java.net.URL;
 
 /**
  * A rule that matches wordy expressions.