+++ languagetool-language-modules/pt/src/main/java/org/languagetool/rules/pt/PortugueseWikipediaRule.java
@@ -23,12 +23,12 @@ import org.languagetool.rules.AbstractSimpleReplaceRule2;
 import org.languagetool.rules.Categories;
 import org.languagetool.rules.Example;
 import org.languagetool.rules.ITSIssueType;
+import org.languagetool.tools.Tools;
 
 import java.io.IOException;
 import java.util.Locale;
 import java.util.ResourceBundle;
 
-import java.net.MalformedURLException;
 import java.net.URL;
 
 /**
@@ -87,11 +87,7 @@ public class PortugueseWikipediaRule extends AbstractSimpleReplaceRule2 {
 
   @Override
   public URL getUrl() {
-    try {
-      return new URL("https://pt.wikipedia.org/wiki/Wikip%C3%A9dia:Lista_de_erros_comuns/M%C3%A1quinas");
-    } catch (MalformedURLException e) {
-      throw new RuntimeException(e);
-    }
+    return Tools.getUrl("https://pt.wikipedia.org/wiki/Wikip%C3%A9dia:Lista_de_erros_comuns/M%C3%A1quinas");
   }
 
   @Override